restructure [component/module] files and fix imports Update all import paths to reflect new structure
