+++ chuchu_scrapper/scrapper.py
@@ -6,7 +6,6 @@ application to extract text, images,
 and links from web pages using HTTP request and HTML parsing.
 """
 
-import argparse
 import os
 import time
 from typing import Dict, List
@@ -14,46 +13,9 @@ from urllib.parse import urlparse
 
 import requests
 
-from modules.myparserlib import ImageParser, LinkParser, TextParser
-
-
-def parse_args():
-    """
-    Parse command line arguments for the ChuChu scraper application.
-
-    Returns:
-        argparse.Namespace: An object containing the parsed command-line arguments.
-
-    The following arguments are supported:
-        - url: URL of the website to scrape
-        - --output: Directory to save the scraped data (default: ./outputs)
-        - --type: Type of data to be scraped (choices: text, image, link, all)
-        - --stdout: Flag to print the scraped data to standard output
-    """
-    parser = argparse.ArgumentParser(
-        prog="chuchu_scrapper",
-        description="A web scraping tool for collecting information from websites.",
-    )
-    parser.add_argument("url", type=str, help="URL of the website to scrape")
-    parser.add_argument(
-        "--output",
-        type=str,
-        default="./outputs",
-        help="Directory to save the scrapped data.",
-    )
-    parser.add_argument(
-        "--type",
-        type=str,
-        choices=["text", "image", "link", "all"],
-        required=True,
-        help="Type of data to be scrapped.",
-    )
-    parser.add_argument(
-        "--stdout",
-        action="store_true",
-        help="Print the scrapped data to standard output.",
-    )
-    return parser.parse_args()
+from chuchu_scrapper.parsers.imageparser import ImageParser
+from chuchu_scrapper.parsers.linkparser import LinkParser
+from chuchu_scrapper.parsers.textparser import TextParser
 
 
 class ChuChuScrapper:
@@ -96,7 +58,7 @@ class ChuChuScrapper:
         textparser.feed(webpage_content)
         return textparser.get_text()
 
-    def scrape_images(self, url) -> List[Dict]:
+    def scrape_images(self, url: str) -> List[Dict]:
         """
         Scrape image data from the given URL.
 
@@ -131,7 +93,7 @@ class ChuChuScrapper:
             return None
 
         parsed = urlparse(url)
-        base_url = f"{parsed.scheme}: //{parsed.netloc}"
+        base_url = {parsed.scheme} + "://" + {parsed.netloc}
 
         linkparser = LinkParser(base_url)
         linkparser.feed(webpage_content)
@@ -271,48 +233,3 @@ class ChuChuScrapper:
         with open(filepath, "w", encoding="utf-8") as f:
             f.write(self.format_links_report(links))
         print(f"Links saved to {filepath}")
-
-
-def main():
-    """
-    Run the ChuChu scraper.
-
-    This function:
-    1. Parses command-line arguments
-    2. Creates a ChuChuScrapper instance
-    3. Executes the appropriate scraping function based on the requested type
-    4. Either prints the results to stdout or saves them to files
-
-    Returns:
-        None
-    """
-    args = parse_args()
-    chuchu_scrapper = ChuChuScrapper()
-
-    if not args.stdout:
-        os.makedirs(args.output, exist_ok=True)
-
-    if args.type == "text":
-        text_content = chuchu_scrapper.scrape_text(args.url)
-        if args.stdout:
-            print("\n=== Scraped Text Content ===\n")
-            print(text_content)
-        else:
-            path = chuchu_scrapper.save_text(text_content, args.output)
-            print(f"Text content saved to {path}")
-
-    elif args.type == "image":
-        image_content = chuchu_scrapper.scrape_images(args.url)
-        if args.stdout:
-            print("\n=== Scraped Image Content ===\n")
-            print(image_content)
-        else:
-            chuchu_scrapper.download_images(image_content, args.output)
-            print(f"Image content saved to {args.output}")
-
-    elif args.type == "link":
-        link_content = chuchu_scrapper.scrape_links(args.url)
-        if args.stdout:
-            print(chuchu_scrapper.format_links_report(link_content))
-        else:
-            chuchu_scrapper.save_links(link_content, args.output)
