+++ pyproject.toml
@@ -19,7 +19,7 @@ requires = ["poetry-core==1.8"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
-main = "chuchu_scrapper.scrapper:main"
+main = "chuchu_scrapper.main:run"
 
 [tool.poetry.dependencies]
 python = ">=3.9,<4.0"
